# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LitXplore",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/review/generate-review": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Generate Review",
        "description": "Start a background task to generate a literature review",
        "operationId": "generateReview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/review/save": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Save Review",
        "operationId": "saveReview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Savereview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/review/history": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Review History",
        "operationId": "getReviewHistory",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Getreviewhistory"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/review/{review_id}": {
      "delete": {
        "tags": [
          "review"
        ],
        "summary": "Delete Review",
        "description": "Delete a specific review",
        "operationId": "deleteReview",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/webhook/clerk": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Clerk Webhook",
        "description": "Handle Clerk webhook events for user synchronization",
        "operationId": "clerk_webhook_api_v1_users_webhook_clerk_post",
        "parameters": [
          {
            "name": "svix-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Svix-Id"
            }
          },
          {
            "name": "svix-timestamp",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Svix-Timestamp"
            }
          },
          {
            "name": "svix-signature",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Svix-Signature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Current User Info",
        "description": "Get the current user's information",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewRequest": {
        "properties": {
          "paper_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Paper Ids"
          },
          "topic": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Topic",
            "description": "Research topic for literature review"
          },
          "max_papers": {
            "type": "integer",
            "maximum": 20,
            "minimum": 1,
            "title": "Max Papers",
            "description": "Maximum number of papers to analyze",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "paper_ids",
          "topic"
        ],
        "title": "ReviewRequest",
        "example": {
          "max_papers": 10,
          "topic": "Recent advances in transformer architectures for natural language processing"
        }
      },
      "SaveReviewRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "description": "Title of the review",
            "default": "Untitled Review"
          },
          "topic": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Topic",
            "description": "Research topic of the review"
          },
          "content": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 1,
            "title": "Content",
            "description": "The review content"
          },
          "citations": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200000
              },
              {
                "type": "null"
              }
            ],
            "title": "Citations",
            "description": "Citations in JSON format"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "content"
        ],
        "title": "SaveReviewRequest",
        "description": "Schema for saving a review with proper input validation."
      },
      "TaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "result_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Data"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at"
        ],
        "title": "TaskResponse"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed"
        ],
        "title": "TaskStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}